Update `Path.join/1` to accept any unicode chardata
We had some issues with the types returned by `Path.join/1`, since it
was technically possible to pass a charlist or IO list to that function
and get something that wasn't a binary, which is what the spec said
should happen. I've updated that to handle charlists and also raw
codepoints.